Add setlist command
Why these changes are being introduced:
For several RDI tasks we need to be able to retrieve the list of sets
from an OAI-PMH repository.

How this addresses that need:
* Changes the main cli command to a click group with two commands,
  harvest and setlist
* Updates both Pipfile and setup.py to include "oai" as the main
  entrypoint command for the cli app
* Cleans up cli.py and adds some logging

Side effects of this change:
This is a backwards-incompatible change to the command structure and
options for this app. From this commit on, other systems using this app
will need to use the updated commands or continue using the 0.x branch
until it is deprecated.