Add workflow to publish to ECR, modify dockerfile and makefile
This commit adds the workflow to automatically publish images on pushes to the main branch.
I update the dockerfile to use the 3.10-slim image, which more closely matches how we do other python dockerfiles.
i update the makefile to have two dev specific commands, "dist-dev" which creates the docker image locally and "publish-dev" which then publishes it to the dev ECR for use in stepfunctions.