Update sentry and logging configuration
Why these changes are being introduced:

Our terraform template requires that each env variable be present in
all envs, even if we don't need it in a particular env (e.g. SENTRY_DSN
is not needed in dev). Env variables are now given a string value of
"None" in env(s) where they are not needed.

In addition, while working on this a logging bug was addressed in which
logging was configured on the cli module logger instead of the root
logger and thus was not propagating to all modules.

In resolving both issues it made sense to refactor these configuration
steps into their own functions in a separate config module.

How this addresses that need:
* Creates a new config module with functions to configure logging and
  sentry, and moves that logic from the cli module to these functions.
* Updates sentry configuration to check for a string value of "None" in
  addition to a None object value for the SENTRY_DSN env variable.
* Adds unit tests for new config module.

Additional changes:
* Minor updates to conftest.py and setup.cfg to set env and caplog
  configuration for all tests.
* Two minor logging additions:
  * One to add a timer to the harvest command and log the total harvest
    time at the end of the process, useful as a benchmark for large/long
    harvest jobs.
  * One to log a status update after every 1000 records harvested,
    useful to monitor progress during large/long harvest jobs.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/RDI-157
* https://mitlibraries.atlassian.net/browse/RDI-127