Allow skip-record and list method option
Why these changes are being introduced:
Because we are setting the skip-record option value via ENV in prod,
this value will always be present even for other sources where we use
the get method. That's fine, harvesting can't skip records from the
wrong source (because they will have different oai-pmh IDs) so we should
allow it. For good measure we can still log a warning that the
skip-record option is being used with the list method, just in case
that's not what someone intends when running this manually or with other
skips/sources in the future.

How this addresses that need:
* Removes sys.exit() from cli harvest command if both skip-record and
  get method options are used.

Side effects of this change:
In theory someone could add record IDs to the skip list, use the list
method, and think those records are being skipped when they won't be.
Since we intend to only use the skip list for records that cause
harvesting errors, this should just result in an exception and the
warning message in the logs will show why.