Gracefully skip records for GetRecords harvest
Why these changes are being introduced:
During a harvest that uses the GetRecords approach, if a particular record throws
and error the harvest fails.  The former workaround was an external skip list, but this
had overhead to maintain.

By gracefully skipping records and reporting them to Sentry, the harvest can continue
even for known problematic records.

How this addresses that need:
* the GetRecords harvest skips errors and sends logs to Sentry
* a new MAX_ALLOWED_ERRORS config value is added to set the upper limit of failed records allowed

Side effects of this change:
* Harvests will complete WITHOUT an external skip list even with problematic records
* DataEng is responsible for monitoring Sentry and triaging skipped records
* We can stop maintaining an external skip list as an SSM parameter

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-258
* https://mitlibraries.atlassian.net/browse/TIMX-257