OAIClient.get_record catch RequestException
Why these changes are being introduced:

A harvest recently failed when the requests library threw a
ConnectTimeout exception.  We were formerly handling only HTTPErrors,
which this exception did not inherit from.  This exception aborted
the harvest instead of gracefully skipping and reporting record,
and continuing.

How this addresses that need:
* OAIClient.get_records() now catches more base RequestException which
includes HTTP and connection errors

Side effects of this change:
* Connection errors when retrieving a record will be logged and
skipped, and will not fail the full harvest.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1038