Remove SENTRY_DSN from required env vars
Why these changes are being introduced:

For local development, or external users of this application, it does not make sense
to require a SENTRY_DSN environment variable.  While setting this to 'none' would
satisfy the config setup, and not use Sentry, it feels unnecessary.

How this addresses that need:
* Moves SENTRY_DSN to optional env vars in Config class

Side effects of this change:
* SENTRY_DSN is no longer required as an env var, but providing it will still get
picked up used

Relevant ticket(s):
* None